// article/src/interfaces/message-enums/seed-logs.ts
export enum SeedLogMessage {
	DATABASE_SEEDING_INIT = 'Database seeding initiated',
	DATABASE_SEEDING_COMPLETED = 'Database seeding completed',
	DATABASE_ALREADY_CONTAINS_DATA = 'Database already contains data. Skipping seeding',
	ERROR_SEEDING_DATABASE = 'Error seeding the database:',
	DATABASE_CLEARING_INITIATED = 'Database clearing initiated',
	DATABASE_CLEARED = 'Database cleared successfully',
	ERROR_CLEARING_DATABASE = 'Error clearing database',
}